requirements.txt
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2==2.9.10
gunicorn==23.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
dj-database-url==2.3.0
requests==2.32.3
